Stable release for campus wifi login with wifi.csv
