--- .vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "python.analysis.autoImportCompletions": true,
-    "python.analysis.typeCheckingMode": "off"
-}
