+++ requirements.txt
@@ -1,1 +1,1 @@
-requests==2.31.0
+requests
