--- wget-log
@@ -1,16 +1,0 @@
-Read error (Connection timed out) in headers.
-Retrying.
-
---2024-09-26 02:33:27--  (try: 2)  https://172.18.10.10:1000/login?
-Connecting to 172.18.10.10:1000... connected.
-WARNING: cannot verify 172.18.10.10's certificate, issued by ‘emailAddress=support@fortinet.com,CN=fortinet-subca2001,OU=Certificate Authority,O=Fortinet,L=Sunnyvale,ST=California,C=US’:
-  Self-signed certificate encountered.
-    WARNING: certificate common name ‘FG10E0TB22901836’ doesn't match requested host name ‘172.18.10.10’.
-HTTP request sent, awaiting response... 200 OK
-Length: 4040 (3.9K) [text/html]
-Saving to: ‘STDOUT’
-
-     0K                                                         0% 0.00 =0s
-
-
-Cannot write to ‘-’ (Success).
